requirements.txt
discord
Pillow
requests
python-dotenv
cogwatch